move reassociate tests
