+++ platform-operator/controllers/verrazzano/component/common/vmi_test.go
@@ -4,6 +4,7 @@
 package common
 
 import (
+	"context"
 	"github.com/stretchr/testify/assert"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
@@ -11,6 +12,8 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	scheme2 "k8s.io/client-go/kubernetes/scheme"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"testing"
 )
@@ -167,4 +170,60 @@ func Test_FindStorageOverride(t *testing.T) {
 			}
 		})
 	}
+}
+
+// TestReassociateResources tests the VMO reassociateResources function
+// GIVEN a VMO component
+//  WHEN I call reassociateResources with a VMO service resource
+//  THEN no error is returned and the VMO service contains expected Helm labels and annotations
+func TestReassociateResources(t *testing.T) {
+	// The actual pre-upgrade testing is performed by the underlying unit tests, this just adds coverage
+	// for the Component interface hook
+	scheme := runtime.NewScheme()
+	_ = corev1.AddToScheme(scheme)
+
+	fakeClient := fake.NewClientBuilder().WithScheme(scheme2.Scheme).WithObjects(&corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: vmoComponentNamespace,
+			Name:      vmoComponentName,
+		},
+	}).Build()
+	err := ExportVMOHelmChart(spi.NewFakeContext(fakeClient, nil, false))
+	assert.NoError(t, err)
+	err = ReassociateVMOResources(spi.NewFakeContext(fakeClient, nil, false))
+	assert.NoError(t, err)
+	service := corev1.Service{}
+	err = fakeClient.Get(context.TODO(), types.NamespacedName{Namespace: vmoComponentNamespace, Name: vmoComponentName}, &service)
+	assert.NoError(t, err)
+	assert.Contains(t, service.Labels["app.kubernetes.io/managed-by"], "Helm")
+	assert.Contains(t, service.Annotations["meta.helm.sh/release-name"], vmoComponentName)
+	assert.Contains(t, service.Annotations["meta.helm.sh/release-namespace"], vmoComponentNamespace)
+	assert.NotContains(t, service.Annotations["helm.sh/resource-policy"], "keep")
+}
+
+// TestExportVmoHelmChart tests the VMO exportVMOHelmChart function
+// GIVEN a VMO component
+//  WHEN I call exportVMOHelmChart with a VMO service resource
+//  THEN no error is returned and the VMO service contains expected Helm labels and annotations
+func TestExportVmoHelmChart(t *testing.T) {
+	// The actual pre-upgrade testing is performed by the underlying unit tests, this just adds coverage
+	// for the Component interface hook
+	scheme := runtime.NewScheme()
+	_ = corev1.AddToScheme(scheme)
+
+	fakeClient := fake.NewClientBuilder().WithScheme(scheme2.Scheme).WithObjects(&corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: vmoComponentNamespace,
+			Name:      vmoComponentName,
+		},
+	}).Build()
+	err := ExportVMOHelmChart(spi.NewFakeContext(fakeClient, nil, false))
+	assert.NoError(t, err)
+	service := corev1.Service{}
+	err = fakeClient.Get(context.TODO(), types.NamespacedName{Namespace: vmoComponentNamespace, Name: vmoComponentName}, &service)
+	assert.NoError(t, err)
+	assert.Contains(t, service.Labels["app.kubernetes.io/managed-by"], "Helm")
+	assert.Contains(t, service.Annotations["meta.helm.sh/release-name"], vmoComponentName)
+	assert.Contains(t, service.Annotations["meta.helm.sh/release-namespace"], vmoComponentNamespace)
+	assert.Contains(t, service.Annotations["helm.sh/resource-policy"], "keep")
 }
